+++ docs/fr_FR/changelog.md
@@ -1,4 +1,7 @@
-# 20/11/2025 (beta)
+# 08/01/2026 (dev)
+- Ajout du n° de ligne dans les logs
+
+# 23/11/2025 (stable)
 - Ajout d'une modale permettant de gérer les tâches planifiées Heatzy (Visualisation, Création, Mise à jour, et suppression)
 
 # 15/11/2025 (stable)